Moved shared package to components package
