+++ app/src/main/java/com/example/deliveryapp/ui/screens/home/HomeScreen.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.deliveryapp.model.Restaurant
 import com.example.deliveryapp.ui.RestaurantItem
-import com.example.deliveryapp.ui.components.menu.shared.BottomAppBar
+import com.example.deliveryapp.ui.components.shared.BottomAppBar
 import kotlinx.serialization.Serializable
 
 @Serializable
